First compilation of ApplicationManager class.
Able to display UI, but need to update main.cpp in order to allow it to
send commands to ApplicationManager. From there, it should be easy to
vet that all features work as previous. Also, UI changes are mostly in
place; all info is at the top now rather than mix of top and bottom.