Improve save/load and misc UI improvements.
Changed how save and load are handled to simply write the ASCII
representation to a text file rather than the numeric code used in the
background. This will make editing scores out of the editor easier.
The UI was changed a bit to test out some new design ideas. Not 100%
sure if it will stay this way or not. Right and left keys also work
with toggling the Index and BPM.