Add support for allowing remote start/stop of play.
This commit adds initial support for having one sequencer control
when other sequencers are to start playing, essentially allowing
for all application instances to commence play at the same moment in
time. UDP Multicast is used to implement this feature; this requires
at least one sequencer to be configured as the "master" with all
others being set up as a "slave".